requirements.txt
numpy
tensorflow
pytest
pillow